Issue #33 : Bug fixes

+++ src/main/java/org/notificationengine/notificator/mail/SingleMailByRecipientNotificator.java
@@ -87,10 +87,14 @@ public class SingleMailByRecipientNotificator extends Notificator {
 
                 Collection<ObjectId> fileIds = (Collection<ObjectId>) context.get("fileIds");
 
-                for(ObjectId fileId : fileIds) {
-                    File file = persister.retrieveFileFromId(fileId);
+                if(fileIds != null) {
 
-                    filesToAttach.add(file);
+                    for(ObjectId fileId : fileIds) {
+                        File file = persister.retrieveFileFromId(fileId);
+
+                        filesToAttach.add(file);
+                    }
+
                 }
 
             }
@@ -121,8 +125,11 @@ public class SingleMailByRecipientNotificator extends Notificator {
 
             //delete files created after sending (or not)
             //even if it has not been sent, the file will be created next time
-            for(File file : filesToAttach) {
-                file.delete();
+            if(filesToAttach != null) {
+
+                for(File file : filesToAttach) {
+                    file.delete();
+                }
             }
 
 		}
